fix[DE]: Extract submenu classes
